🚧Update kd-admin to latest version (#397)
* Start work for issue #393

* fix: update kd-admin version in install scripts

updates the kd-admin version in the install scripts to
v1.0.0-preview.8 to pick up the latest changes.

* config: update schema urls for dev tools settings

updates the schema urls in the create-pr, prepare-release, and
gen-release-notes settings files to point to the 'preview' branch
instead of a specific tag. this ensures that the settings always
use the latest schema definitions from the preview branch.

* config: map issue categories to issue types in release notes

maps issue categories to corresponding issue types in both the
previous and production release notes settings files, and removes
duplicate category label mappings. this provides a more structured
and organized approach to categorizing issues in release notes.